requirements.txt
Flask==3.1.0
flask-cors==5.0.1
PyInstaller==6.3.0
pywebview==4.4